requirements.txt
flask
gunicorn
requests
BeautifulSoup4
PyYAML==6.0.1
